Move the default stores to OpenIddict.Stores and the server components to OpenIddict.Server
