+++ src/OpenIddict.EntityFramework/Stores/OpenIddictTokenStore.cs
@@ -16,6 +16,7 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.Caching.Memory;
 using OpenIddict.Core;
 using OpenIddict.Models;
+using OpenIddict.Stores;
 
 namespace OpenIddict.EntityFramework
 {
