+++ src/OpenIddict.Server/Internal/OpenIddictInitializer.cs
