+++ src/OpenIddict.Server/Internal/OpenIddictProvider.Authentication.cs
@@ -5,7 +5,6 @@
  */
 
 using System;
-using System.Collections.Immutable;
 using System.IO;
 using System.Threading.Tasks;
 using AspNet.Security.OpenIdConnect.Extensions;
