+++ src/OpenIddict.Server/Internal/OpenIddictProvider.Discovery.cs
