+++ src/OpenIddict.Server/Internal/OpenIddictProvider.Helpers.cs
