+++ src/OpenIddict.Server/Internal/OpenIddictProvider.Serialization.cs
