+++ src/OpenIddict.Stores/OpenIddict.Stores.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="..\..\build\packages.props" />
 
@@ -7,14 +7,14 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <Description>Easy-to-use OpenID Connect server for ASP.NET Core.</Description>
+    <Description>Default base stores for OpenIddict.</Description>
     <Authors>Kévin Chalet</Authors>
     <PackageTags>aspnetcore;authentication;jwt;openidconnect;openiddict;security</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\OpenIddict.Server\OpenIddict.Server.csproj" />
-    <ProjectReference Include="..\OpenIddict.Stores\OpenIddict.Stores.csproj" />
+    <ProjectReference Include="..\OpenIddict.Core\OpenIddict.Core.csproj" />
+    <ProjectReference Include="..\OpenIddict.Models\OpenIddict.Models.csproj" />
   </ItemGroup>
 
   <ItemGroup>
