+++ src/OpenIddict.Stores/Stores/OpenIddictScopeStore.cs
@@ -14,9 +14,10 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.Caching.Memory;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using OpenIddict.Core;
 using OpenIddict.Models;
 
-namespace OpenIddict.Core
+namespace OpenIddict.Stores
 {
     /// <summary>
     /// Provides methods allowing to manage the scopes stored in a database.
