+++ test/OpenIddict.EntityFramework.Tests/OpenIddictExtensionsTests.cs
@@ -21,7 +21,10 @@ namespace OpenIddict.EntityFrameworkCore.Tests
             // Arrange
             var builder = new OpenIddictBuilder(new ServiceCollection())
             {
-                ApplicationType = typeof(object)
+                ApplicationType = typeof(object),
+                AuthorizationType = typeof(OpenIddictAuthorization),
+                ScopeType = typeof(OpenIddictScope),
+                TokenType = typeof(OpenIddictToken)
             };
 
             // Act and assert
@@ -40,7 +43,10 @@ namespace OpenIddict.EntityFrameworkCore.Tests
             // Arrange
             var builder = new OpenIddictBuilder(new ServiceCollection())
             {
-                AuthorizationType = typeof(object)
+                ApplicationType = typeof(OpenIddictApplication),
+                AuthorizationType = typeof(object),
+                ScopeType = typeof(OpenIddictScope),
+                TokenType = typeof(OpenIddictToken)
             };
 
             // Act and assert
@@ -59,7 +65,10 @@ namespace OpenIddict.EntityFrameworkCore.Tests
             // Arrange
             var builder = new OpenIddictBuilder(new ServiceCollection())
             {
-                ScopeType = typeof(object)
+                ApplicationType = typeof(OpenIddictApplication),
+                AuthorizationType = typeof(OpenIddictAuthorization),
+                ScopeType = typeof(object),
+                TokenType = typeof(OpenIddictToken)
             };
 
             // Act and assert
@@ -78,6 +87,9 @@ namespace OpenIddict.EntityFrameworkCore.Tests
             // Arrange
             var builder = new OpenIddictBuilder(new ServiceCollection())
             {
+                ApplicationType = typeof(OpenIddictApplication),
+                AuthorizationType = typeof(OpenIddictAuthorization),
+                ScopeType = typeof(OpenIddictScope),
                 TokenType = typeof(object)
             };
 
@@ -100,7 +112,13 @@ namespace OpenIddict.EntityFrameworkCore.Tests
         {
             // Arrange
             var services = new ServiceCollection();
-            var builder = new OpenIddictBuilder(services);
+            var builder = new OpenIddictBuilder(services)
+            {
+                ApplicationType = typeof(OpenIddictApplication),
+                AuthorizationType = typeof(OpenIddictAuthorization),
+                ScopeType = typeof(OpenIddictScope),
+                TokenType = typeof(OpenIddictToken)
+            };
 
             // Act
             builder.AddEntityFrameworkStores<DbContext>();
