+++ test/OpenIddict.Server.Tests/Internal/OpenIddictProviderTests.Authentication.cs
