+++ test/OpenIddict.Server.Tests/Internal/OpenIddictProviderTests.Discovery.cs
