+++ test/OpenIddict.Server.Tests/Internal/OpenIddictProviderTests.Serialization.cs
